pyproject.toml
[project]
name = "OptiFeed"
version = "0.1.0"
description = "A Telegram bot for financial news and stock market analysis, powered by Google Generative AI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.13",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "pika>=1.3.2",
    "py2neo>=2021.2.4",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.64",
]
